Possible fix for the production error/introducing magic session timeout
Possible fix for production error: try longer to find the node (20 times
500ms rather than 3). 

Introduction of magic number for session timeout. If the session timeout
is 1 there will be no session storage. Redis will not be used.

Removed some code that should no longer be used: the old way of message
crypto is now removed.